+++ Gruntfile.js
@@ -4,32 +4,31 @@ module.exports = function(grunt) {
 	// Read project configuration
 
 	var  path = require('path')
-		,configPath = '_config/'
-		,configFiles = [
-			 'default'
-			,'project'
-			,'secret'
-			,'local'
-		]
-		,cfg ,modulesPattern ,buildBaseDir
+		,configr = require('./lib/configure')
+		,cfg ,modulesPattern, buildBasePath, buildBaseDirName
+		,isDistBuild = grunt.option('dist') || false // based on this value we will execute a dev or dist build
 	;
 
+
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-	// Read test configuration (if testing)
+	// Get project configuration
 
-	// Override project config, if command line options contain a config location.
-	configPath = grunt.option('config-path') || configPath;
-	configFiles = grunt.option('config-files') && grunt.option('config-files').split(',') || configFiles;
-
-	// Merge configurations
-	cfg = require('./lib/configure').merge(configPath, configFiles).get();
+	// Override project config, if command line options contain config options (for testing)
+	cfg = grunt.option('config-path') && grunt.option('config-files')
+		? configr.merge(grunt.option('config-path'), grunt.option('config-files').split(','))
+		: configr.get()
+	;
 
 
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 	// Create vars from config
 
 	modulesPattern = path.join(cfg.paths.modulesBaseDir, cfg.moduleDirName.replace('{{name}}', '*'))
-	buildBaseDir = path.join(cfg.paths.staticBaseDir, cfg.static.build.baseDirName); // build.baseDirName may be empty
+	buildBaseDirName = isDistBuild ? cfg.static.build.baseDirNameDist : cfg.static.build.baseDirNameDev;
+	buildBasePath = isDistBuild
+		? path.join(cfg.paths.staticBaseDir, cfg.static.build.baseDirNameDist)  // build.baseDirNameDist can be empty
+		: path.join(cfg.paths.staticBaseDir, cfg.static.build.baseDirNameDev)   // build.baseDirNameDev can be empty
+	;
 
 
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -37,38 +36,42 @@ module.exports = function(grunt) {
 
 	grunt.initConfig({
 
-		pkg: grunt.file.readJSON('package.json')
+		pkg                         : grunt.file.readJSON('package.json')
 
-		,staticUriPrefix        : (cfg.staticUriPrefix || '') + '/'
-		,staticUriPrefixCss     : cfg.staticUriPrefix
+		,staticUriPrefixCss         : cfg.staticUriPrefix
+		,staticUriPrefix            : (cfg.staticUriPrefix || '') + '/'
 
 
 		////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 		// Sources base paths
 
-		,tcInline               : cfg.paths.inline
-		,tcBase                 : cfg.paths.base
-		,tcModules              : modulesPattern
-		,tcApplication          : cfg.paths.application
-		,staticDir              : cfg.paths.staticBaseDir
+		,tcInline                   : cfg.paths.inline
+		,tcBase                     : cfg.paths.base
+		,spritesPath                : '<%=tcBase%>/css/sprites'
+		,tcModules                  : modulesPattern
+		,tcApplication              : cfg.paths.application
+		,staticDir                  : cfg.paths.staticBaseDir
 
 
 		////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 		// Output destinations
 
-		,baseDirName            : cfg.static.build.baseDirName
-		,buildBaseDir           : buildBaseDir
-		,tmp                    : '<%=buildBaseDir%>/tmp'
-		,destJs                 : '<%=buildBaseDir%>'
-		,destCss                : '<%=buildBaseDir%>'
-		,destSpritesCss         : '<%=tcBase%>/css/sprites' // technically this should go in tmp, but we want the generated classes in our base css for easy lookup.
-		,destSpritesImg         : '<%=buildBaseDir%>'
+		,baseDirName                : cfg.static.build.baseDirName
+		
+		,buildBasePath              : buildBasePath             // dynamic, depends on build mode (development/production)
+		,buildBaseDirName           : buildBaseDirName          // dynamic, depends on build mode (development/production)
+		,tmpPath                    : '<%=buildBasePath%>/tmp'
+
+		,destJs                     : '<%=buildBasePath%>'
+		,destCss                    : '<%=buildBasePath%>'
+		,destSpritesImg             : '<%=buildBasePath%>'
+		,destSpritesCss             : '<%=spritesPath%>'        // technically this should go in tmp, but we want the generated classes in our base css for easy lookup.
 
 
 		////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 		// Directory names
 
-		,skinsDirName           : cfg.skinsDirName
+		,skinsDirName               : cfg.skinsDirName
 
 
 		////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -82,7 +85,7 @@ module.exports = function(grunt) {
 			,external_css: [
 				 '<%=tcBase%>/css/sprites/*.less'
 				,'<%=tcBase%>/css/lib/*.{less,css}'
-				,'<%=tcBase%>/css/elements/*.less'
+				,'<%=tcBase%>/css/*.less'
 				,'<%=tcModules%>/*.less'
 				,'<%=tcModules%>/<%=skinsDirName%>/*.less'
 				,'<%=tcApplication%>/css/*.less'
@@ -114,9 +117,9 @@ module.exports = function(grunt) {
 				 '<%=staticDir%>/_static/test/*.js'
 				,'<%=tcModules%>/test/*.js'
 			]
-			,sprites: [
-				'<%=tcBase%>/css/sprites/'
-				//,'<%=tcModules%>/sprites/' // not implemented
+			,sprites_watch: [
+				 '<%=spritesPath%>/**/*.{png|jpg|conf}'
+				//,'<%=tcModules%>/sprites/*.{png|jpg|conf}' // not implemented yet
 			]
 		}
 
@@ -125,40 +128,51 @@ module.exports = function(grunt) {
 		// Configure Grunt plugins
 
 		,glue: {
-			icons: {
-				src: '<%=sources.sprites%>'
-				//,dest: '<%=destSpritesCss%>/00-sprites.less' // not implemented
-				,options: '--css=<%=destSpritesCss%> --img=<%=destSpritesImg%> --less --url=<%=staticUriPrefix%><%=baseDirName%> --namespace=s --sprite-namespace= --recursive --crop --optipng --force --debug'
+			// see: https://github.com/MarcDiethelm/grunt-glue-nu
+			options: {
+				 css                : '<%=destSpritesCss%>'
+				,less               : true
+				,url                : '<%=staticUriPrefix%><%=buildBaseDirName%>'
+				,namespace          : ''
+				,'sprite-namespace' : ''
+				,optipng            : true
+			}
+			,sprites: {
+				options: {
+					'sprite-namespace' : 's' // set to inline-block in sprites-helper.less
+				}
+				,src                : ['<%=spritesPath%>/misc', '<%=tcModules%>/sprites']
+				,dest               : '<%=buildBasePath%>'
 			}
 		}
 
 		,less_imports: {
 			inline: {
-				 src: '<%=sources.inline_css%>'
-				,dest: '<%=tmp%>/inline-imports.less'
+				 src                : '<%=sources.inline_css%>'
+				,dest               : '<%=tmpPath%>/inline-imports.less'
 			},
 			external: {
 				options: {}
-				,src: '<%=sources.external_css%>'
-				,dest: '<%=tmp%>/external-imports.less'
+				,src                : '<%=sources.external_css%>'
+				,dest               : '<%=tmpPath%>/external-imports.less'
 			}
 		}
 
 		,concat: {
 			inline_scripts: {
 				options: {
-					banner: '/*! Inline script dependencies for page bootstrapping */\n'
+					banner          : '/*! Inline script dependencies for page bootstrapping */\n'
 				}
-				,src: '<%=sources.inline_js%>'
-				,dest: '<%=destJs%>/inline.js'
+				,src                : '<%=sources.inline_js%>'
+				,dest               : '<%=destJs%>/inline.js'
 			}
 			,external_scripts: {
-				 src: '<%=sources.external_js%>'
-				,dest: '<%=destJs%>/external.js'
+				 src                : '<%=sources.external_js%>'
+				,dest               : '<%=destJs%>/external.js'
 			}
 			,module_tests: {
-				 src: '<%=sources.module_test_js%>'
-				,dest: '<%=destJs%>/test.js'
+				 src                : '<%=sources.module_test_js%>'
+				,dest               : '<%=destJs%>/test.js'
 			}
 		}
 
@@ -167,79 +181,91 @@ module.exports = function(grunt) {
 				options: {
 					preserveComments: 'some'
 				}
-				,src: '<%=destJs%>/inline.js'
-				,dest: '<%=destJs%>/inline.min.js'
+				,src                : '<%=destJs%>/inline.js'
+				,dest               : '<%=destJs%>/inline.min.js'
 			},
 			external: {
-				 src: '<%=destJs%>/external.js'
-				,dest: '<%=destJs%>/external.min.js'
+				 src                : '<%=destJs%>/external.js'
+				,dest               : '<%=destJs%>/external.min.js'
 			}
 		}
 
 		,less: {
 			inline: {
 				options: {
-					 // cssmin will not create file if the output is empty. a special comment fixes this.
-					 banner: '/*! Inline style dependencies for page bootstrapping */'
+					                // cssmin will not create file if the output is empty. a special comment fixes this.
+					 banner         : '/*! Inline style dependencies for page bootstrapping */'
 					,imports: {
-						reference: ['<%=tcInline%>/css/lib/reference/*.less']
+						reference   : ['<%=tcInline%>/css/lib/reference/*.less']
 					}
 				}
-				,src: '<%=tmp%>/inline-imports.less'
-				,dest: '<%=destCss%>/inline.css'
+				,src                : '<%=tmpPath%>/inline-imports.less'
+				,dest               : '<%=destCss%>/inline.css'
 			}
 			,external: {
 				options: {
-					 banner: "@static-prefix: '<%=staticUriPrefixCss%>';"
+					 banner         : "@static-prefix: '<%=staticUriPrefixCss%>';"
 					,imports: {
-						reference: ['<%=tcBase%>/css/lib/reference/*.less']
+						reference   : ['<%=tcBase%>/css/lib/reference/*.less']
 					}
 				}
-				,src: '<%=tmp%>/external-imports.less'
-				,dest: '<%=destCss%>/external.css'
+				,src                : '<%=tmpPath%>/external-imports.less'
+				,dest               : '<%=destCss%>/external.css'
 			}
 		}
 
 		,cssmin: {
 			inline: {
-				 src: '<%=destCss%>/inline.css'
-				,dest: '<%=destCss%>/inline.min.css'
+				 src                : '<%=destCss%>/inline.css'
+				,dest               : '<%=destCss%>/inline.min.css'
 			},
 			external: {
-				 src: '<%=destCss%>/external.css'
-				,dest: '<%=destCss%>/external.min.css'
+				 src                : '<%=destCss%>/external.css'
+				,dest               : '<%=destCss%>/external.min.css'
 			}
+		}
+
+		// After dist build remove all non-minified files.
+		,clean: {
+			dist: [
+				 '<%=tmpPath%>'
+				,'<%=destCss%>/inline.css'
+				,'<%=destCss%>/external.css'
+				,'<%=destJs%>/inline.js'
+				,'<%=destJs%>/external.js'
+			]
 		}
 
 		,jshint: {
 			options: {
 				// report but don't fail
-				force: true
+				force               : true
 				// enforce
-				,latedef: true
-				,undef: true
+				,latedef            : true
+				,undef              : true
 				//relax
-				,laxcomma: true
-				,smarttabs: true
-				,expr: true
-				,asi: true
-				,loopfunc: true
-				,nonew: true
+				,laxcomma           : true
+				,smarttabs          : true
+				,expr               : true
+				,asi                : true
+				,loopfunc           : true
+				,nonew              : true
 				// environment
-				,browser: true
-				,jquery: true
+				,browser            : true
+				,jquery             : true
 				,globals: {
-					$: true
-					,Tc: true
-					,xtc: true
+					$               : true
+					,Tc             : true
+					,xtc            : true
+					,console        : true
 					// Lawg
-					,log: true
-					,dir: true
-					,info: true
-					,debug: true
-					,warn: true
-					,error: true
-					,table: true
+					,log            : true
+					,dir            : true
+					,info           : true
+					,debug          : true
+					,warn           : true
+					,error          : true
+					,table          : true
 					// QUnit ಠ_ಠ
 					,asyncTest      : true
 					,deepEqual      : true
@@ -258,35 +284,52 @@ module.exports = function(grunt) {
 					,test           : true
 				}
 			}
-			,inline: ['<%=sources.jshint_inline%>']
-			,external: ['<%=sources.jshint_external%>']
-			,module_tests: ['<%=sources.jshint_module_test%>']
+			,inline                 : ['<%=sources.jshint_inline%>']
+			,external               : ['<%=sources.jshint_external%>']
+			,module_tests           : ['<%=sources.jshint_module_test%>']
 		}
 
 		,watch: {
 			sprites: {
-				files: ['<%=sources.sprites%>*.{png,jpg}']
-				,tasks: ['build-sprites', 'build-external-css']
+				 files              : ['<%=sources.sprites_watch%>']
+				,tasks              : ['build-sprites', 'build-external-css']
 			},
 			inline_styles: {
-				files: ['<%=sources.inline_css%>', '<%=tcInline%>/css/lib/reference/*.less']
-				,tasks: ['build-inline-css']
+				 files              : ['<%=sources.inline_css%>', '<%=tcInline%>/css/lib/reference.less']
+				,tasks              : ['build-inline-css']
 			},
 			external_styles: {
-				files: ['<%=sources.external_css%>', '<%=tcBase%>/css/lib/reference/*.less']
-				,tasks: ['build-external-css']
+				 files              : ['<%=sources.external_css%>', '<%=tcBase%>/css/lib/reference.less']
+				,tasks              : ['build-external-css']
 			}
 			,inline_scripts: {
-				files: ['<%=sources.inline_js%>']
-				,tasks: ['build-inline-js']
+				 files              : ['<%=sources.inline_js%>']
+				,tasks              : ['build-inline-js']
 			}
 			,external_scripts: {
-				files: ['<%=sources.external_js%>']
-				,tasks: ['build-external-js']
+				 files              : ['<%=sources.external_js%>']
+				,tasks              : ['build-external-js']
 			}
 			,module_tests: {
-				files: ['<%=sources.module_test_js%>']
-				,tasks: ['build-module-tests']
+				 files              : ['<%=sources.module_test_js%>']
+				,tasks              : ['build-module-tests']
+			}
+			,gruntfile: {
+				 files              : ['Gruntfile.js']
+				,tasks              : 'default'
+			}
+		}
+
+		,log: {
+			options: {
+				 bold: true
+				,color: 'green'
+			}
+			,devCssDone: {
+				msg: 'Dev CSS complete'
+			}
+			,devJsDone: {
+				msg: 'Dev JS complete'
 			}
 		}
 	});
@@ -296,17 +339,17 @@ module.exports = function(grunt) {
 	// Define build tasks                            // use actual task name (first part before colon)!
 
 	grunt.registerTask('build-sprites',              ['glue']);
-	grunt.registerTask('build-inline-js',            ['concat:inline_scripts', 'uglify:inline', 'jshint:inline']);
-	grunt.registerTask('build-external-js',          ['concat:external_scripts', 'uglify:external', 'jshint:external',]);
-	grunt.registerTask('build-inline-css',           ['less_imports:inline',   'less:inline',   'cssmin:inline']);
-	grunt.registerTask('build-external-css',         ['less_imports:external', 'less:external', 'cssmin:external']);
+	grunt.registerTask('build-external-js',          ['concat:external_scripts', 'jshint:external',]);
+	grunt.registerTask('build-inline-js',            ['concat:inline_scripts', 'log:devJsDone', 'jshint:inline']);
+	grunt.registerTask('build-external-css',         ['less_imports:external', 'less:external']);
+	grunt.registerTask('build-inline-css',           ['less_imports:inline', 'less:inline', 'log:devCssDone']);
 	grunt.registerTask('build-module-tests',         ['concat:module_tests' , 'jshint:module_tests']);
 
 
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-	// Define a default task
+	// Define the build pipes
 
-	var defaultTask = [
+	var devBuild = [
 		 'build-sprites'
 		,'build-external-css'
 		,'build-external-js'
@@ -316,15 +359,32 @@ module.exports = function(grunt) {
 		,'watch'
 	];
 
+	var distBuild = [
+		 'build-sprites'
+		,'build-external-css'
+		,'build-external-js'
+		,'build-inline-css'
+		,'build-inline-js'
+		,'build-module-tests'
+
+		,'uglify:inline'
+		,'uglify:external'
+		,'cssmin:inline'
+		,'cssmin:external'
+		,'clean:dist'
+	];
+
+	var task = isDistBuild ? distBuild : devBuild;
+
 	// If sprites building is not enabled in the app config, remove it.
-	!cfg.enableSpritesBuilding && defaultTask.shift();
-	grunt.registerTask('default', defaultTask);
+	!cfg.enableSpritesBuilding && task.shift();
+	grunt.registerTask('default', task);
 
 
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 	// Load grunt plugins
 
-	grunt.loadNpmTasks('grunt-glue');
+	grunt.loadNpmTasks('grunt-glue-nu');
 	grunt.loadNpmTasks('grunt-less-imports');
 	grunt.loadNpmTasks('assemble-less');
 	grunt.loadNpmTasks('grunt-contrib-concat');
@@ -332,5 +392,23 @@ module.exports = function(grunt) {
 	grunt.loadNpmTasks('grunt-contrib-jshint');
 	grunt.loadNpmTasks('grunt-contrib-cssmin');
 	grunt.loadNpmTasks('grunt-contrib-watch');
+	grunt.loadNpmTasks('grunt-contrib-clean');
+
+
+	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+	// Ad-hoc plugin
+
+	grunt.registerMultiTask('log', 'a console logging task', function() {
+		var options = this.options()
+			,color = options.color || false
+		;
+
+		if (options.bold) {
+			grunt.log.subhead(this.data.msg[color]);
+		}
+		else {
+			grunt.log.writeln(this.data.msg[color]);
+		}
+	});
 
 };