+++ _config/config-local.js
@@ -1,8 +1,0 @@
-// Never commit authentication data to a public repository!
-// config-local.js listed in .gitignore
-// config-local.js listed in .jitsuignore
-
-// copy any config property you need and override it here for local development
-// just be sure it's not part of your deployments
-module.exports = {
-}