+++ _config/config-project.js
@@ -1,3 +1,0 @@
-// Never commit authentication data to a public repository!
-module.exports = {
-}