+++ controllers/_default.js
@@ -1,17 +1,18 @@
 module.exports = function(app) {
 
-	var  cfg = app.cfg
+	var  cfg = require('../lib/configure').get()
 		,path = require('path')
-		,utils = require(path.join(cfg.dirname, '/lib/utils.js'))
 		,fs = require('fs')
+
+		,renderModule = require('../lib/mod-render')()
+		,exphbs = require('express3-handlebars')
 		,utils = require('../lib/utils')
-		,docTitle = app.docTitle
-		// for useful error message when an asset is not found
-		,assetUriPrefix = cfg.staticUriPrefix + '/' + cfg.static.build.baseDirName +'/'
-		,cssUri = assetUriPrefix + cfg.static.build.css.external.development
-		,jsUri  = assetUriPrefix + cfg.static.build.js.external.development
-		,assetsRegExp = new RegExp(cssUri +'|'+ jsUri);
-	;
+		,helpers = require('../lib/helpers.js')
+		,docTitle = helpers.docTitle
+		,hbs
+	 ;
+
+	hbs = exphbs.create({ handlebars: require('handlebars') });
 
 	return {
 
@@ -20,38 +21,38 @@ module.exports = function(app) {
 		 */
 
 		_home: function(req, res, next) {
-			var overview = require(path.join(cfg.dirname, 'lib/overview.js'))(cfg);
+			var overview = require(path.join(cfg.appPath, 'lib/overview.js'))(cfg);
 
 			res.render('_app-overview', {
 				 layout: false
 				,docTitle: docTitle('Components Overview')
 				,title: 'Components Overview'
 				,views: overview.views
-				,templates: overview.templates
+				,layouts: overview.layouts
 				,modules: overview.modules
 			});
 		}
 
 		,_getView: function(req, res, next) {
-			var layout = cfg.defaultTemplateName;
+			var layout = cfg.defaultLayoutName;
 
 			if ('raw' in req.query) {
 				res.setHeader('Content-Type', 'text/plain');
 				layout = false;
 			}
-			res.render(req.params.name, {
+			res.render(req.params.view, {
 				 layout: layout
-				,docTitle: docTitle('View: '+ req.params.name)
+				,docTitle: docTitle('View: '+ req.params.view)
 				,uri: req.originalUrl
 				,skipModules: 'solo' in req.query && 'layout'
 			});
 		}
 
 		,_getModule: function(req, res, next) {
-			var module = app.terrific.renderModule(
+			var module = renderModule(
 					app.locals,
 					{
-						 name: req.params.name
+						 name: req.params.module
 						,template: req.params.template
 					}
 				)
@@ -68,14 +69,14 @@ module.exports = function(app) {
 			} else {
 				res.locals(app.locals);
 				res.locals({
-					layout: false
-					,docTitle: docTitle('Module: '+ req.params.name +', Template: '+ req.params.template)
+					 layout: false
+					,docTitle: docTitle('Module: '+ req.params.module +', Template: '+ req.params.template)
 					,body: module
-					,exclusive: req.params.name
+					,exclusive: req.params.module
 					,skipModules: true
 				});
 
-				app.hbs.render(path.join(cfg.paths.templates, cfg.defaultTemplateName + '.hbs'), res.locals,
+				hbs.render(path.join(cfg.paths.layouts, cfg.defaultLayoutName + '.hbs'), res.locals,
 					function(err, html) {
 						if (err) {
 							var error = utils.error('Unable to render the module in the default template', err);
@@ -88,42 +89,42 @@ module.exports = function(app) {
 			}
 		}
 
-		,_getTemplate: function(req, res, next) {
+		,_getLayout: function(req, res, next) {
 			res.locals(app.locals);
 			res.locals({
 				layout: false
-				,docTitle: docTitle('Template: '+ req.params.name)
+				,docTitle: docTitle('Layout: '+ req.params.layout)
 				,body: ''
 			});
 
-			app.hbs.render(path.join(cfg.paths.templates, req.params.name + '.hbs'), res.locals,
+			hbs.render(path.join(cfg.paths.layouts, req.params.layout + '.hbs'), res.locals,
 				function(err, html) {
 					if (err) {
-						var error = utils.error('Unable to render the template', err);
+						var error = utils.error('Unable to render the layout', err);
 						console.error(error.c);
 						html = error.web;
 					}
 
 					'raw' in req.query
 						&& res.setHeader('Content-Type', 'text/plain');
-					
+
 					res.send(html);
 				}
 			);
 		}
 
 		,_getModuleTest: function(req, res, next) {
-			var test = req.query.url;
-
-			var output = ''
+			var  test = req.query.url
+				,output = ''
 				,modules = []
 				,done = {}// keep track of handled module, to skip modules with identical options.
 			;
+
 			app.tests[test].forEach(function(options) {
 				if (options.name+options.template+options.skins+options.connectors in done)
 					return;
-				
-				output += app.terrific.renderModule(app.locals, options);
+
+				output += renderModule(app.locals, options);
 				modules.push({
 					 name:      options.name
 					,template:  options.template
@@ -135,19 +136,20 @@ module.exports = function(app) {
 
 			res.locals(app.locals);
 			res.locals({
-				layout: false
+				 layout: false
 				,body: '<script>var xtc = {isTest: true, testModules: '+ JSON.stringify(modules) +'}</script>\n\n'+ output
 				 // Prevent initializing testing in the test frame by overwriting the handlebars helper 'test'.
 				,helpers: { test: null }
 			});
 
-			app.hbs.render(path.join(cfg.pathsAbsolute.templates, cfg.defaultTemplateName + '.hbs'), res.locals,
+			hbs.render(path.join(cfg.pathsAbsolute.layouts, cfg.defaultLayoutName + '.hbs'), res.locals,
 				function(err, html) {
 					if (err) {
 						var error = utils.error('Unable to render the modules in the default template', err);
 						console.error(error.c);
 						html = error.web;
 					}
+
 					res.send(html);
 				}
 			);
@@ -156,11 +158,11 @@ module.exports = function(app) {
 
 		 // Look for a view with the name supplied by the catch-all route
 		,_subPage: function(req, res, next) {
-			fs.exists(path.join(cfg.pathsAbsolute.views, req.params.pageName + '.hbs'), function(exists) {
-				if ( exists ) {
+			fs.exists(path.join(cfg.pathsAbsolute.views, req.params.view + '.hbs'), function(exists) {
+				if (exists) {
 					try {
-						res.render(req.params.pageName, {
-							 docTitle: docTitle(req.params.pageName)
+						res.render(req.params.view, {
+							 docTitle: docTitle('View: '+ req.params.view)
 							,uri: req.originalUrl
 						});
 					} catch (e) {
@@ -168,40 +170,9 @@ module.exports = function(app) {
 					}
 				}
 				else {
-					return app.render404(req, res, next);
+					return helpers.render404(req, res, next);
 				}
 			});
-		}
-
-		/////////////////////////////////////////////////////////////////////
-
-		// If no Express middleware sends a response before, this middleware is finally called.
-		,render404: function(req, res, next) {
-
-			var isAssetUri = assetsRegExp.test(req.url)
-				,err
-			;
-
-			if (isAssetUri) {
-				err = utils.error('404 NOT FOUND '+ req.url, null, '==> Did you forget to run grunt?');
-				console.error( err.c );
-				res
-					.type('text/plain')
-					.send(404, err.web)
-			;
-			}
-			else {
-				res
-					.status(404)
-					.render(
-					'404'
-					,{
-						 docTitle: docTitle('404')
-						,uri: req.originalUrl
-					}
-				);
-			}
-
 		}
 	}
 };