+++ test/_config/config-test.js
@@ -1,4 +1,0 @@
-// Config for all tests, overrides defaults
-module.exports = {
-	devPort: 3333
-}